LPD-25338 Move to correct folder
